Reworked views: removed redundant html tags
This commit moves unnecessary  and redudant tags (e.g. a .col-md-12 in a .row or a .row within a .row) A further reduction of tag-use in the individual views was achieved by moving the common container divs into _Layout.

This also fixes some responsive design issues in the admin view.